Refactor: remove Py* structs, add mashumaro, move scripts to nzcvm.scripts with Tap, replace CoordinateSystem
Agent-Logs-Url: https://github.com/lispandfound/nzcvm/sessions/a9543f5a-4169-4ed0-800f-82c78e3008c6

+++ src/model_tree.rs
@@ -129,7 +129,7 @@ impl Query for ModelTree {
             simplex_tests: outer_stats.simplex_tests,
             hit_count,
             output: quality,
-            elapsed: now.elapsed().as_nanos(),
+            elapsed: now.elapsed().as_nanos() as u64,
         }
     }
